+++ apps/frontend/src/shared/api/index.ts
@@ -1,0 +1,2 @@
+export * from './kakao'
+export * from './room'
