+++ apps/frontend/src/shared/api/kakao/index.ts
@@ -1,0 +1,1 @@
+export * from './kakao'
