+++ apps/frontend/src/shared/api/room/index.ts
@@ -1,0 +1,1 @@
+export * from './room'
