+++ apps/frontend/src/shared/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useKakaoQueries'
-export * from './useRoomQueries'
-export * from './useSocketClient'
-export * from './useToast'
+export * from './queries'
+export * from './socket'
+export * from './toast'
