+++ apps/frontend/src/shared/hooks/socket/index.ts
@@ -1,0 +1,1 @@
+export * from './useSocketClient'
