+++ apps/frontend/src/shared/hooks/socket/useSocketClient.ts
