+++ apps/frontend/src/shared/hooks/toast/index.ts
@@ -1,0 +1,1 @@
+export * from './useToast'
